test(default_enum_value): Move tests to separate module
- Move related tests to separate module
- Check default of `CMsgRemoteClientBroadcastHeader`
- Make config dependency explicit in `build.rs`
